🎯 Attach and perform middlewares
